+++ src/main/java/springboot/dao/UserDao.java
@@ -7,12 +7,12 @@ import java.util.List;
 public interface UserDao {
     List<User> allUser();
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
     void addUser(User user);
 
     void updateUser(User user);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 
 }
